Proyecto/requirements.txt
# Librerías y dependencias para el proyecto MLOps
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
jupyter==1.0.0
notebook==6.5.4
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
fastapi==0.100.0
uvicorn==0.23.2
pydantic==1.10.13
joblib==1.3.2
mlflow==2.5.0
evidently==0.2.8
psutil==5.9.5
scipy
streamlit
lightgbm
colorama
Pillow==10.0.0